+++ InvokeBuildHelper/Functions/Get-IBHConfig.ps1
@@ -37,6 +37,10 @@ function Get-IBHConfig
             Enabled             = $true
         }
 
+        SchemaTask          = [PSCustomObject] @{
+            TextFileExtension   = '.gitignore', '.gitattributes', '.ps1', '.psm1', '.psd1', '.ps1xml', '.txt', '.xml', '.cmd', '.json', '.md'
+        }
+
         AnalyzeTask         = [PSCustomObject] @{
             ScriptAnalyzerRules = Get-ScriptAnalyzerRule
         }