+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Added: New task LocalDebug added for deployment of a module to the local system
 * Added: Schema test for the .debug.ps1 file
 * Added: Schema tests for the (helper) function files and export definitions
+* Added: Schema tests for encoding, white space and indentation character
 * Changed: Move asserts from Gallery to Approve task
 
 ## 1.0.0 - 2019-10-23